Updated GameInfo.java to hold more information for the Game.
Listens for category change.
Listens for letter change.

Checks if the letter is in the word.
First it checks if the category is solved and if not then checks the first word, if the first word is solved then second then third. CheckLetterClicked() checks if the letter is in the word and if it is then guess left stay as it is and if the letter is not in the word than guess left decrease.