Added listView to Client GUI
I added a listview to the client gui so that the client can get updated with game info, such as if the letter they selected was in the word or not, as well as reporting back the number of guesses remaining and other game info.

In addition, I also added a setOnCloseRequest in the server gui so that when the user closes the server, it frees up the port that it was using and doesnt cause issues.